Files for the backend work of logging users in
login.py contains code that validates user login information, and provides them access to the actual tool if they provide a username and password pair that is on record. users.csv is the csv file containing all valid login information at present. In order for this code to be fully operational, a proper path to the csv file should be provided.